requirements.txt
pyyaml>=3.12
pykwalify>=1.5.1
pypandoc>=1.1.3
tox-travis>=0.5
tox>=2.3.1
pip>8.1.0
wheel>0.25.0
